`fn mask_edges_intra`: Make `sidx` vars `usize`s.

+++ src/lf_mask.rs
@@ -289,10 +289,10 @@ unsafe fn mask_edges_intra(
     let mut mask = 1 << by4;
     y = 0;
     while y < h4 {
-        let sidx = (mask >= 0x10000) as libc::c_int;
+        let sidx = (mask >= 0x10000) as usize;
         let smask = mask >> (sidx << 4);
-        let ref mut fresh4 = masks[0][bx4 as usize]
-            [imin(twl4c, l[y as usize] as libc::c_int) as usize][sidx as usize];
+        let ref mut fresh4 =
+            masks[0][bx4 as usize][imin(twl4c, l[y as usize] as libc::c_int) as usize][sidx];
         *fresh4 = (*fresh4 as libc::c_uint | smask) as u16;
         y += 1;
         mask <<= 1;
@@ -300,10 +300,10 @@ unsafe fn mask_edges_intra(
     x = 0;
     mask = 1 << bx4;
     while x < w4 {
-        let sidx = (mask >= 0x10000) as libc::c_int;
+        let sidx = (mask >= 0x10000) as usize;
         let smask = mask >> (sidx << 4);
-        let ref mut fresh5 = masks[1][by4 as usize]
-            [imin(thl4c, a[x as usize] as libc::c_int) as usize][sidx as usize];
+        let ref mut fresh5 =
+            masks[1][by4 as usize][imin(thl4c, a[x as usize] as libc::c_int) as usize][sidx];
         *fresh5 = (*fresh5 as libc::c_uint | smask) as u16;
         x += 1;
         mask <<= 1;
